requirements.txt
streamlit
pandas
numpy
python-docx
scikit-learn
PyPDF2
reportlab
